Refactor parser into package
