+++ src/main/java/Hamlet.java
@@ -1,6 +1,7 @@
 package main.java;
 
 import main.java.exception.HamletException;
+import main.java.parser.Parser;
 import main.java.storage.Storage;
 import main.java.task.TaskList;
 import main.java.ui.Ui;
